+++ Composite Primary Key JPA Demo/composite-primary-key-jpa/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/jpa_composite_pk_demo1
 spring.datasource.username=postgres
 spring.datasource.password=admin1234
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
